# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "zen API",
    "version": "1.0.0"
  },
  "paths": {
    "/cards": {
      "get": {
        "summary": "List Cards",
        "description": "The mind's board, newest first. `after` is the id of the last card already seen: its own\n(created_at, id) is the cursor, so equal timestamps never skip or repeat a card.",
        "operationId": "list_cards_api_cards_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "exclusiveMinimum": 0,
              "default": 60,
              "title": "Limit"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "title": "After",
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Card",
        "description": "Capture returns the saved card at once, with no model call; enrichment (tags, embedding,\nthe daily mark unlocks) is a background job from a later ticket.",
        "operationId": "create_card_api_cards_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardInput",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}": {
      "patch": {
        "summary": "Edit Card",
        "description": "Edit, pin, mark seen, restore, or set the resurfacing date; only the fields the request set\nare touched.",
        "operationId": "edit_card_api_cards__id__patch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardPatch",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Let Card Go",
        "description": "A soft delete: `PATCH` with `restore: true` brings the card back as it was.",
        "operationId": "let_card_go_api_cards__id__delete",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}/article": {
      "get": {
        "summary": "Read Article",
        "description": "Reading Mode's saved copy of a card this mind owns, sanitised again on every read (the\nsame defence-in-depth the TypeScript server applies) rather than trusted from storage alone.",
        "operationId": "read_article_api_cards__id__article_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Article": {
        "properties": {
          "article_html": {
            "type": "string",
            "title": "Article Html"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "article_html"
        ],
        "title": "Article",
        "description": "Reading Mode's saved copy, sanitised again on every read (defence in depth)."
      },
      "Card": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "note",
              "quote",
              "link",
              "article",
              "image",
              "video",
              "product",
              "book",
              "movie",
              "recipe",
              "tweet",
              "person",
              "color",
              "font",
              "pdf",
              "file"
            ],
            "title": "Kind"
          },
          "title": {
            "title": "Title",
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "title": "Url",
            "type": [
              "string",
              "null"
            ]
          },
          "domain": {
            "title": "Domain",
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "title": "Note",
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "title": "Content",
            "type": [
              "string",
              "null"
            ]
          },
          "image_path": {
            "title": "Image Path",
            "type": [
              "string",
              "null"
            ]
          },
          "meta": {
            "additionalProperties": true,
            "propertyNames": {
              "type": "string"
            },
            "type": "object",
            "title": "Meta"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "has_article": {
            "type": "boolean",
            "title": "Has Article"
          },
          "pinned_at": {
            "title": "Pinned At",
            "type": [
              "string",
              "null"
            ]
          },
          "seen_at": {
            "title": "Seen At",
            "type": [
              "string",
              "null"
            ]
          },
          "enriched_at": {
            "title": "Enriched At",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "url",
          "domain",
          "note",
          "content",
          "image_path",
          "meta",
          "colors",
          "tags",
          "has_article",
          "pinned_at",
          "seen_at",
          "enriched_at",
          "created_at",
          "updated_at"
        ],
        "title": "Card",
        "description": "A saved card: a note, link, image or other kind of memory."
      },
      "CardInput": {
        "properties": {
          "url": {
            "title": "Url",
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "title": "Note",
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "title": "Title",
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "title": "Kind",
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "note",
              "quote",
              "link",
              "article",
              "image",
              "video",
              "product",
              "book",
              "movie",
              "recipe",
              "tweet",
              "person",
              "color",
              "font",
              "pdf",
              "file"
            ]
          },
          "quote": {
            "title": "Quote",
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "title": "Content",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CardInput",
        "description": "A new card: give a url, a note, a quote, or any mix."
      },
      "CardPatch": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "note",
              "quote",
              "link",
              "article",
              "image",
              "video",
              "product",
              "book",
              "movie",
              "recipe",
              "tweet",
              "person",
              "color",
              "font",
              "pdf",
              "file"
            ]
          },
          "title": {
            "title": "Title",
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "title": "Note",
            "type": [
              "string",
              "null"
            ]
          },
          "colors": {
            "title": "Colors",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "tags": {
            "title": "Tags",
            "items": {
              "type": "string",
              "maxLength": 60
            },
            "type": [
              "array",
              "null"
            ]
          },
          "pinned": {
            "title": "Pinned",
            "type": [
              "boolean",
              "null"
            ]
          },
          "seen": {
            "title": "Seen",
            "type": [
              "boolean",
              "null"
            ]
          },
          "restore": {
            "title": "Restore",
            "type": [
              "boolean",
              "null"
            ]
          },
          "resurface": {
            "title": "Resurface",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CardPatch",
        "description": "A partial edit: only the fields present in the request are changed."
      },
      "CardResponse": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "card"
        ],
        "title": "CardResponse"
      },
      "CardsList": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "type": "array",
            "title": "Cards"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cards"
        ],
        "title": "CardsList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "Ok"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Supabase access token, verified against the project's published signing keys."
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}